cloud_probe/orchestrator.py
import boto3
from azure.mgmt.monitor import MonitorManagementClient
from google.cloud import monitoring_v3
from typing import Dict, Any
import os

class CloudOrchestrator:
    """
    Orchestrates proactive failover across AWS, Azure, and GCP.
    In a production environment, this would interact with DNS (Route53/CloudDNS), 
    Load Balancers, and Traffic Managers.
    """
    
    def __init__(self):
        # Initializing clients (would use env vars for credentials in production)
        self.aws_region = os.getenv("AWS_REGION", "us-east-1")
        # Clients would be initialized here if credentials were present
        # self.aws_client = boto3.client('route53')
        
    async def trigger_aws_failover(self) -> Dict[str, str]:
        """
        Example: Update Route53 health checks or weights to shift traffic.
        """
        # Mocking AWS SDK call
        return {"cloud": "AWS", "status": "SUCCESS", "action": "Traffic shifted via Route53"}

    async def trigger_azure_failover(self) -> Dict[str, str]:
        """
        Example: Update Azure Traffic Manager profiles.
        """
        # Mocking Azure SDK call
        return {"cloud": "Azure", "status": "SUCCESS", "action": "Traffic Manager profile updated"}

    async def trigger_gcp_failover(self) -> Dict[str, str]:
        """
        Example: Update Cloud DNS or Global Load Balancer.
        """
        # Mocking GCP SDK call
        return {"cloud": "GCP", "status": "SUCCESS", "action": "Global Load Balancer updated"}

    async def execute_multi_cloud_failover(self) -> Dict[str, Any]:
        """
        Triggers failover across all providers in parallel.
        """
        # For simplicity in this example, we execute them sequentially
        results = [
            await self.trigger_aws_failover(),
            await self.trigger_azure_failover(),
            await self.trigger_gcp_failover()
        ]
        
        return {
            "overall_status": "COMPLETED",
            "provider_results": results
        }
